review: address minor PR comments on sign-tx foundation
- Drop signTxClientFactory var indirection — SignAndBroadcast now calls
  newSDKTxClient inline; validation guards moved into lowercase
  signAndBroadcast so tests can dispatch directly with a fake txClient.
- Split SDK txClient (sdkTxClient + newSDKTxClient + the three methods +
  newSignTxInterfaceRegistry + makeSignTxCodec) into transactions.go,
  leaving sign_and_broadcast.go as the orchestration layer.
- Keep rpc.NodeInfo.Network naming — that's the CometBFT protocol term;
  comment now explains the cross-check with task params.chainId.

Follow-up #171 tracks swapping these local rpc types for upstream
sei-tendermint/rpc/coretypes (touches peers/statesync/shadow, so out
of scope for #170).
